--- src/image/mod.rs
@@ -1,1 +1,0 @@
-pub mod maimai_best_50;
