+++ src/service/mod.rs
@@ -1,2 +1,5 @@
+pub mod maimai_best_50;
+pub mod printer;
 pub mod resource;
+pub mod simple_log;
 pub mod table;
