+++ src/service/printer.rs
