+++ src/service/simple_log.rs
