+++ src/utils/mod.rs
@@ -1,4 +1,2 @@
 pub mod file;
-pub mod printer;
-pub mod simple_log;
-pub mod image;
+pub(crate) mod image;
