+++ tests/command_test.rs
